+++ src/main/java/com/jfeat/am/module/order/api/OrderItemRewardEndpoint.java
@@ -63,7 +63,7 @@ public class OrderItemRewardEndpoint {
 
     @PutMapping("/{id}")
     @ApiOperation(value = "修改 OrderItemReward", response = OrderItemReward.class)
-    public Tip updateOrderItemReward(@PathVariable Integer id, @RequestBody OrderItemReward entity) {
+    public Tip updateOrderItemReward(@PathVariable Long id, @RequestBody OrderItemReward entity) {
         entity.setId(id);
         return SuccessTip.create(orderItemRewardService.updateMaster(entity));
     }
@@ -107,7 +107,7 @@ public class OrderItemRewardEndpoint {
                                      @RequestParam(name = "pageNum", required = false, defaultValue = "1") Integer pageNum,
                                      @RequestParam(name = "pageSize", required = false, defaultValue = "10") Integer pageSize,
                                      @RequestParam(name = "search", required = false) String search,
-                                     @RequestParam(name = "id", required = false) Integer id,
+                                     @RequestParam(name = "id", required = false) Long id,
                                      @RequestParam(name = "orderId", required = false) Integer orderId,
                                      @RequestParam(name = "orderNumber", required = false) String orderNumber,
                                      @RequestParam(name = "orderTotalPrice", required = false) BigDecimal orderTotalPrice,
